channel teaser component, use on channels page
