+++ client/src/data/channels.json
@@ -11,6 +11,7 @@
     {
       "title": "Test Channel",
       "slug": "test-channel",
+      "teaser": "Hello teaser.",
       "description": "Hello description."
     }
   ]
