+++ client/src/routes/Channels.tsx
@@ -2,22 +2,19 @@ import React, { Component } from 'react'
 import Route from '../components/templates/Route'
 import Content from '../components/atoms/Content'
 import channels from '../data/channels.json'
-import { Link } from 'react-router-dom'
+import ChannelTeaser from '../components/organisms/ChannelTeaser'
 
 class Channels extends Component {
     public render() {
         return (
             <Route title={channels.title} description={channels.description}>
-                <Content>
-                    <ul>
-                        {channels.items.map(channel => (
-                            <li key={channel.title}>
-                                <Link to={`/channels/${channel.slug}`}>
-                                    {channel.title}
-                                </Link>
-                            </li>
-                        ))}
-                    </ul>
+                <Content wide>
+                    {channels.items.map(channel => (
+                        <ChannelTeaser
+                            key={channel.title}
+                            channel={channel.slug}
+                        />
+                    ))}
                 </Content>
             </Route>
         )
